Refactor CompileTools to remove anything vscode related
